# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Swagger TS3",
    "description": "TS3 API",
    "contact": {
      "email": "contact@n1eko.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://tsapi.n1eko.com/api"
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "Operations about users"
    },
    {
      "name": "Channels",
      "description": "Operations about channels"
    },
    {
      "name": "Auth",
      "description": "Operations about authentication"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users connected",
        "description": "Get all users connected",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "content": {}
          }
        }
      }
    },
    "/channels/{id}": {
      "get": {
        "tags": [
          "Channels"
        ],
        "summary": "Get channel info",
        "description": "Get channel info",
        "operationId": "getChannelInfo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Channel ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          },
          "500": {
            "description": "Invalid channel ID",
            "content": {}
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register user",
        "description": "Register user to authenticate in API",
        "operationId": "register",
        "security": [
          {
            "register-token": []
          }
        ],
        "requestBody": {
          "description": "User object to be registered",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "user"
                  ],
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {}
          },
          "401": {
            "description": "Register token missing or invalid",
            "content": {}
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login user",
        "description": "Login user to authenticate in API",
        "operationId": "login",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "User object to be logged in",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserToken"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {}
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "register-token": {
        "type": "apiKey",
        "name": "register-token",
        "in": "header"
      }
    },
    "schemas": {
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "isMuted": {
            "type": "boolean"
          },
          "country": {
            "type": "string"
          }
        }
      },
      "Channel": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "format": "int64"
          },
          "isSecured": {
            "type": "boolean"
          }
        }
      },
      "UserLogin": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "UserToken": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      }
    }
  }
}